fix naive debugger line numbers in ServerErrorMiddleware (#821)
Previously the error middleware in debug mode would naively calculate the line
numbers without knowledge of the position in the file. This meant that if the
exception occurred near the top or bottom of the file, the line numbers and
the highlighted line would be incorrect. The named tuple returned by
`inspect.getinnerframes` contains an `index` field which is not naive and will
result in correct line numbers. This change also simplifies the code somewhat